refactor : 기존 RequestByUseCase안의 UseCase 별 Request를 개별로 모두 분리
