+++ K-Market/K-Market/Source/Data/ProductStorage/Infra/Request/DetailFetchRequest.swift
@@ -1,0 +1,18 @@
+//
+//  DetailFetchRequest.swift
+//  K-Market
+//
+//  Created by parkhyo on 2023/03/18.
+//
+
+import Foundation
+
+struct DetailFetchRequest: CustomRequest {
+    var path: String?
+    var query: [URLQueryItem]?
+    var httpMethod: HTTPMethod = .GET
+    
+    init(id: Int) {
+        self.path = "/api/products/\(id)"
+    }
+}
