+++ K-Market/K-Market/Source/Data/ProductStorage/Infra/Request/ListFetchRequest.swift
@@ -1,0 +1,21 @@
+//
+//  ListFetchRequest.swift
+//  K-Market
+//
+//  Created by parkhyo on 2023/03/18.
+//
+
+import Foundation
+
+struct ListFetchRequest: CustomRequest {
+    var path: String? = "/api/products"
+    var query: [URLQueryItem]?
+    var httpMethod: HTTPMethod = .GET
+    
+    init(pageNo: Int, itemsPerPage: Int) {
+        query = [
+            URLQueryItem(name: "page_no", value: String(pageNo)),
+            URLQueryItem(name: "items_per_page", value: String(itemsPerPage))
+        ]
+    }
+}
